requirements.txt
evo
matplotlib
numpy
opencv-python
pandas
pillow
pyarrow
pybind11
pyparsing
pyyaml
python-math
rasterio
scikit-image
scikit-learn
scipy==1.14.1
tqdm
cartopy
geopy
